backend/requirements.txt
fastapi
uvicorn
python-jose
passlib
python-multipart
pandas
numpy
scikit-learn
joblib
pydantic
pydantic-settings